requirements.txt
google-generativeai==0.8.4
pyinstaller==6.11.1
pyinstaller-hooks-contrib>=2024.0 